requirements.txt
alembic==0.9.9
attrs==19.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
blinker==1.4
click==6.7
colorama==0.3.9
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
enum34==1.1.6
Flask==0.12.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.3.2
future==0.16.0
gitdb2==2.0.3
GitPython==2.1.9
gunicorn==19.7.1
inflection==0.3.1
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==3.0.1
jupyter-core==4.4.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
pandas==0.23.4
pandocfilters==1.4.2
Pygments==2.2.0
pyrsistent==0.14.11
python-dateutil==2.7.2
python-editor==1.0.3
PyYAML==3.12
six==1.11.0
smmap2==2.0.3
soupsieve==1.9
SQLAlchemy==1.2.6
tabulate==0.8.2
testpath==0.4.2
toml==0.9.4
traitlets==4.3.2
webencodings==0.5.1
Werkzeug==0.14.1
yapf==0.26.0
